Add some organization to the transition examples.
Grouping the examples together by related functionality and feature.
